Move Navigator to runtime
