+++ circuit-codegen/src/main/kotlin/com/slack/circuit/codegen/CircuitSymbolProcessorProvider.kt
@@ -59,7 +59,7 @@ private val CIRCUIT_UI = ClassName(CIRCUIT_BASE_PACKAGE, "Ui")
 private val CIRCUIT_UI_FACTORY = CIRCUIT_UI.nestedClass("Factory")
 private val CIRCUIT_UI_STATE = ClassName(CIRCUIT_RUNTIME_BASE_PACKAGE, "CircuitUiState")
 private val SCREEN = ClassName(CIRCUIT_RUNTIME_BASE_PACKAGE, "Screen")
-private val NAVIGATOR = ClassName(CIRCUIT_BASE_PACKAGE, "Navigator")
+private val NAVIGATOR = ClassName(CIRCUIT_RUNTIME_BASE_PACKAGE, "Navigator")
 private val CIRCUIT_CONTEXT = ClassName(CIRCUIT_RUNTIME_BASE_PACKAGE, "CircuitContext")
 private const val FACTORY = "Factory"
