+++ circuit/src/jvmTest/kotlin/com/slack/circuit/EventListenerTest.kt
@@ -12,6 +12,7 @@ import app.cash.turbine.Turbine
 import com.google.common.truth.Truth.assertThat
 import com.slack.circuit.runtime.CircuitContext
 import com.slack.circuit.runtime.CircuitUiState
+import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
 import java.time.LocalTime
 import java.time.ZoneOffset.UTC
