+++ samples/interop/src/main/kotlin/com/slack/circuit/sample/interop/MainActivity.kt
@@ -32,7 +32,7 @@ import com.slack.circuit.CircuitCompositionLocals
 import com.slack.circuit.CircuitConfig
 import com.slack.circuit.CircuitContent
 import com.slack.circuit.runtime.CircuitContext
-import com.slack.circuit.Navigator
+import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.Presenter
 import com.slack.circuit.Ui
 import com.slack.circuit.sample.counter.CounterPresenterFactory
