+++ samples/star/src/main/kotlin/com/slack/circuit/star/home/HomeScreen.kt
@@ -19,7 +19,7 @@ import com.slack.circuit.CircuitContent
 import com.slack.circuit.runtime.CircuitUiEvent
 import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.NavEvent
-import com.slack.circuit.Navigator
+import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.onNavEvent
