+++ samples/star/src/main/kotlin/com/slack/circuit/star/navigator/AndroidSupportingNavigator.kt
@@ -3,7 +3,7 @@
 package com.slack.circuit.star.navigator
 
 import android.content.Intent
-import com.slack.circuit.Navigator
+import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.Screen
 import kotlinx.parcelize.Parcelize
