+++ samples/star/src/main/kotlin/com/slack/circuit/star/petlist/PetList.kt
@@ -66,7 +66,7 @@ import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.slack.circuit.runtime.CircuitUiEvent
 import com.slack.circuit.runtime.CircuitUiState
-import com.slack.circuit.Navigator
+import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.Presenter
 import com.slack.circuit.runtime.Screen
 import com.slack.circuit.codegen.annotations.CircuitInject
